pyproject.toml
[tool.poetry]
authors = ["hemengyang <hmy0119@gmail.com>"]
description = "基于 NoneBot2 的聊天机器人"
license = "MIT"
name = "coolqbot"
readme = "README.md"
repository = "https://github.com/he0119/CoolQBot"
version = "0.15.3"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = { extras = ["httpx", "websockets"], version = "^2.0.0-beta.4" }
nonebot-adapter-onebot = "^2.1.0"
nonebot-adapter-qqguild = "^0.1.0"
nonebot-plugin-apscheduler = "^0.1.2"
nonebot-plugin-sentry = "^0.2.2"
nonebot-plugin-datastore = "^0.3.1"
nonebot-plugin-wordcloud = "0.2.3"
tencentcloud-sdk-python = "^3.0.593"
psutil = "^5.8.0"
nonebot-plugin-memes = "0.3.3"
nonebot-plugin-petpet = "0.3.6"
nonebot-bison = "0.5.4"
nonebot-plugin-treehelp = "^0.0.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
nonebug = "^0.2.0"
pytest-cov = "^3.0.0"
pytest-mock = "^3.6.1"
pytest-xdist = "^2.5.0"
pre-commit = "^2.19.0"

[tool.nonebot]
plugin_dirs = ["src/plugins"]
plugins = [
    "nonebot_plugin_apscheduler",
    "nonebot_plugin_sentry",
    "nonebot_plugin_treehelp",
    "nonebot_plugin_datastore",
    "nonebot_plugin_wordcloud",
    "nonebot_plugin_memes",
    "nonebot_plugin_petpet",
    "nonebot_bison",
]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88
skip_gitignore = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
